% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{xbrlBaseElement}
\alias{xbrlBaseElement}
\title{Base Element Details}
\usage{
xbrlBaseElement(Element, Namespace = NULL, as_data_frame = TRUE)
}
\arguments{
\item{Element}{The element name in the base taxonomy}

\item{Namespace}{(optional) The namespace of the taxonomy the data is requested for.
For example http://fasb.org/us-gaap/2015-01-31}

\item{as_data_frame}{Returns a data frame (TRUE by default)}
}
\description{
Gets details of US GAAP Taxonomy element
}
\examples{
\dontrun{
xbrlBaseElement("Assets")
}
}
\references{
\url{https://github.com/xbrlus/data_analysis_toolkit/blob/master/api/xbrlBaseElement.md}
}

