% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{xbrlChildren}
\alias{xbrlChildren}
\title{Children}
\usage{
xbrlChildren(Element, AccessionID = NULL, GroupURI, Linkbase = NULL,
  Accession = NULL, NetworkLink = NULL, as_data_frame = TRUE)
}
\arguments{
\item{Element}{The element name in the base taxonomy.
This parameter will not take a comma separated list.}

\item{AccessionID}{Internal Accession identifier used by the XBRL US database.
This is a unique filing identifier.
For example one company will have many filings.
This is returned by the API and can be used in subsequent calls.
This allows a comma separated list}

\item{GroupURI}{The extended link role in an XBRL report that is defined by
the company.}

\item{Linkbase}{The type of network relationship. This could be a
Presentation, Calculation or Definition.
If this is not entered then all will be returned.}

\item{Accession}{Filing accession number.
This is the accession number used as the filing identifier used by the SEC.
This parameter does not allow a comma separated list.}

\item{NetworkLink}{The relationship type that links two elements together.
For example summation-item in the calculation linkbase or dimension-default
in the definition linkbase.}

\item{as_data_frame}{Return value in a data frame (default)}
}
\description{
Gets the relationships in a network by passing the extended link role,
an element name and the filing number/CIK.
}
\details{
All calls to the API must include the Element parameter name and
  at least an AccessionID or an Accession number.
  In addition the extended link role must be reported.
}
\examples{
\dontrun{
  xbrlChildren(
    Element = "Assets", AccessionID = 120617,
    GroupURI = "http://www.thecocacolacompany.com/role/ConsolidatedBalanceSheets",
    Linkbase = "Calculation", NetworkLink = "summation-item")
  )
}
}
\references{
\url{https://github.com/xbrlus/data_analysis_toolkit/blob/master/api/xbrlChildren.md}
}

