% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{xbrlCIKLookup}
\alias{xbrlCIKLookup}
\title{CIK Lookup}
\usage{
xbrlCIKLookup(Ticker, as_data_frame = TRUE)
}
\arguments{
\item{Ticker}{The ticker of the company}

\item{as_data_frame}{Returns a data frame (TRUE by default)}
}
\description{
Gets Central Index Key and other information about a company from a ticker symbol
}
\examples{
\dontrun{
xbrlCIKLookup("aapl")
}
}
\references{
\url{https://github.com/xbrlus/data_analysis_toolkit/blob/master/api/xbrlCIKLookup.md}
}

