% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{xbrlNetwork}
\alias{xbrlNetwork}
\title{Network}
\usage{
xbrlNetwork(Element, Linkbase = NULL, AccessionID = NULL,
  Accession = NULL, CIK = NULL)
}
\arguments{
\item{Element}{The element name in the base taxonomy.
This parameter will not take a comma separated list.}

\item{Linkbase}{The type of network relationship. This could be a
Presentation, Calculation or Definition.
If this is not entered then all will be returned.}

\item{AccessionID}{Internal Accession identifier used by the XBRL US database.}

\item{Accession}{Filing accession number.
This is the accession number used as the filing identifier used by the SEC.}

\item{CIK}{CIK of the Company. This must be 10 digits in length.
This parameter allows a comma separated list.}
}
\description{
Gets the relationships in a network by passing the extended link role,
an element name and the filing number/CIK.
}
\details{
This API allows the user to fetch details about a report
  (Group/Network/Extended link role) in an XBRL filing that an element
  appears in such as the balance sheet or income statement.
  This could return multiple results as Assets for example could be in
  multiple locations in a filing.
  The user passes an element and filing number/CIK and the report url will be
  returned.
  The API allows the user to specify the different linkbases associated with
  a report. For example a user can request the calculation network for those
  reports that contain Assets for company ABC.

  All calls to the API must include the Element parameter name and at least
  an AccessionID or an Accession number or a CIK parameter.
}
\examples{
\dontrun{
  xbrlNetwork(Element="Assets", AccessionID = 103575)
}
}
\references{
\url{https://github.com/xbrlus/data_analysis_toolkit/blob/master/api/xbrlNetwork.md}
}

