% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbrlus.R
\docType{package}
\name{xbrlus-package}
\alias{xbrlus}
\alias{xbrlus-package}
\title{xbrlus: Interface to XBRL US API}
\description{
This package provides an interface to XBRL US API.
}
\details{
All APIs (except for the CIKLookup) require use of a valid XBRL US API
key. You can get the key and read the terms of usage at
\url{http://xbrl.us/use/howto/data-analysis-toolkit/}.

\pkg{xbrlus} functions will read the API key from environment variable
\code{XBRLUS_API_KEY}.
To start R session with the initialized environvent variable
create an .Renviron file in your R home with a line like this:

\code{XBRLUS_API_KEY=************}

To check where your R home is, try \code{normalizePath("~")}.
}
\references{
\itemize{
   \item{Data Analysis Toolkit and API description on GitHub: \url{https://github.com/xbrlus/data_analysis_toolkit}}
 }
}

