% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{xbrlExtensionElement}
\alias{xbrlExtensionElement}
\title{Extension Element}
\usage{
xbrlExtensionElement(Element, AccessionID = NULL, Accession = NULL,
  Namespace = NULL, as_data_frame = TRUE)
}
\arguments{
\item{Element}{The element name in the base taxonomy.
This parameter will not take a comma separated list.}

\item{AccessionID}{Internal Accession identifier used by the XBRL US database.}

\item{Accession}{Filing accession number.
This is the accession number used as the filing identifier used by the SEC.}

\item{Namespace}{The namespace of the company filing the data is requested for.
For example http://www.ovt.com/20150430.}

\item{as_data_frame}{Return value in a data frame (default)}
}
\description{
Gets details about elements used in the company extensions.
}
\details{
This API allows the user to fetch details of about elements used in
  the company extensions in an XML format, by passing the element name,
  namespace and entity information in the API.
  THe API allows the user to get the attributes of an extension element and
  the associated labels.
  If the filing information is provided the API will return the labels used
  by the company in their extension filing.
  As a convenience it will also return the attributes of the US GAAP taxonomy.
}
\examples{
\dontrun{
  xbrlExtensionElement(
    AccessionID = 103575,
    Element = "ResearchDevelopmentAndRelatedExpenses"
  )
}
}
\references{
\url{https://github.com/xbrlus/data_analysis_toolkit/blob/master/api/xbrlExtensionElement.md}
}

